+++ concrete_datastore/concrete/automation/tasks.py
@@ -23,6 +23,9 @@ from concrete_datastore.concrete.models import (  # pylint:disable=E0611
 from concrete_datastore.api.v1.permissions import (
     update_created_by_permissions,
     create_or_update_instance_permission_per_user,
+    check_instance_permissions_per_user,
+    bulk_create_permission_instances,
+    bulk_update_permission_instances,
 )
 
 
@@ -34,8 +37,20 @@ DIVIDER_MODEL_LOWER = DIVIDER_MODEL.lower()
 
 
 @app.task
+def check_all_user_permissions_async(user_pk, new_level=None):
+    user = get_user_model().objects.get(pk=user_pk)
+    instances_to_create, instances_to_update = check_instance_permissions_per_user(
+        user=user, user_level=new_level
+    )
+    bulk_create_permission_instances(instances_to_create)
+    bulk_update_permission_instances(instances_to_update)
+
+
+@app.task
 def on_update_divider_async(pk_set, instance_uid, include_divider):
     divider_model = getattr(concrete_datastore.concrete.models, DIVIDER_MODEL)
+    instances_to_create = []
+    instances_to_update = []
     for meta_model in meta_models:
         model_name = meta_model.get_model_name()
         if model_name in ('User', 'Group', 'Email'):
@@ -47,13 +62,21 @@ def on_update_divider_async(pk_set, instance_uid, include_divider):
         related_divider_field_name = f'divider_{related_name}s'
         for divider_pk in pk_set:
             divider_instance = divider_model.objects.get(pk=divider_pk)
-            create_or_update_instance_permission_per_user(
+            instance, should_create = create_or_update_instance_permission_per_user(
                 user=get_user_model().objects.get(pk=instance_uid),
                 instances_qs=getattr(
                     divider_instance, related_divider_field_name
                 ).all(),
                 include_divider=include_divider,
             )
+            if instance is None:
+                continue
+            if should_create is True:
+                instances_to_create.append(instance)
+            else:
+                instances_to_update.append(instance)
+    bulk_create_permission_instances(instances_to_create)
+    bulk_update_permission_instances(instances_to_update)
 
 
 @app.task
@@ -65,6 +88,8 @@ def on_update_group_members_async(
         concrete_datastore.concrete.models, instance_model_name
     )
     instance = instance_model.objects.get(pk=instance_uid)
+    instances_to_create = []
+    instances_to_update = []
     for meta_model in meta_models:
         model_name = meta_model.get_model_name()
         if model_name in ('User', 'Group', 'Email'):
@@ -79,12 +104,18 @@ def on_update_group_members_async(
             administrable_qs = getattr(instance, group_admin_field_name).all()
             for user_pk in pk_set:
                 user = user_model.objects.get(pk=user_pk)
-                create_or_update_instance_permission_per_user(
+                perm_instance, should_create = create_or_update_instance_permission_per_user(
                     user=user,
                     instances_qs=viewable_qs | administrable_qs,
                     include_admin_groups=include_pks,
                     include_view_groups=include_pks,
                 )
+                if perm_instance is None:
+                    continue
+                if should_create is True:
+                    instances_to_create.append(perm_instance)
+                else:
+                    instances_to_update.append(perm_instance)
         else:
             #: Instance is User
             user_viewable_field_name = f'viewable_{related_model_name}s'
@@ -97,18 +128,28 @@ def on_update_group_members_async(
             instances_qs = concrete_model.objects.filter(
                 Q(can_admin_groups__in=pk_set) | Q(can_view_groups__in=pk_set)
             )
-            create_or_update_instance_permission_per_user(
+            perm_instance, should_create = create_or_update_instance_permission_per_user(
                 user=instance,
                 instances_qs=instances_qs,
                 include_view_users=include_pks,
                 include_admin_users=include_pks,
             )
+            if perm_instance is None:
+                continue
+            if should_create is True:
+                instances_to_create.append(perm_instance)
+            else:
+                instances_to_update.append(perm_instance)
+    bulk_create_permission_instances(instances_to_create)
+    bulk_update_permission_instances(instances_to_update)
 
 
 @app.task
 def on_create_instance_async(user_pk, model_name, instance_pk):
     model = getattr(concrete_datastore.concrete.models, model_name)
     instance = model.objects.get(pk=instance_pk)
+    instances_to_create = []
+    instances_to_update = []
     if user_pk is not None:
         user = get_user_model().objects.get(pk=user_pk)
         if user.is_at_least_admin is False:
@@ -125,9 +166,17 @@ def on_create_instance_async(user_pk, model_name, instance_pk):
     ):
         if user.is_at_least_admin:
             continue
-        create_or_update_instance_permission_per_user(
+        instance, should_create = create_or_update_instance_permission_per_user(
             user=user, instances_qs=model.objects.filter(pk=instance_pk)
         )
+        if instance is None:
+            continue
+        if should_create is True:
+            instances_to_create.append(instance)
+        else:
+            instances_to_update.append(instance)
+    bulk_create_permission_instances(instances_to_create)
+    bulk_update_permission_instances(instances_to_update)
 
 
 @app.task
@@ -141,13 +190,23 @@ def on_view_admin_groups_changed_async(
         .distinct()
     )
     model = getattr(concrete_datastore.concrete.models, model_name)
+    instances_to_create = []
+    instances_to_update = []
     for user in user_model.objects.filter(pk__in=users_ids):
-        create_or_update_instance_permission_per_user(
+        instance, should_create = create_or_update_instance_permission_per_user(
             user=user,
             instances_qs=model.objects.filter(pk=instance_pk),
             include_view_groups=include_view_groups,
             include_admin_groups=include_admin_groups,
         )
+        if instance is None:
+            continue
+        if should_create is True:
+            instances_to_create.append(instance)
+        else:
+            instances_to_update.append(instance)
+    bulk_create_permission_instances(instances_to_create)
+    bulk_update_permission_instances(instances_to_update)
 
 
 @app.task
@@ -156,13 +215,23 @@ def on_view_admin_users_changed_async(
 ):
     user_model = get_user_model()
     model = getattr(concrete_datastore.concrete.models, model_name)
+    instances_to_create = []
+    instances_to_update = []
     for user in user_model.objects.filter(pk__in=pk_set):
-        create_or_update_instance_permission_per_user(
+        instance, should_create = create_or_update_instance_permission_per_user(
             user=user,
             instances_qs=model.objects.filter(pk=instance_pk),
             include_admin_users=include_admin_users,
             include_view_users=include_view_users,
         )
+        if instance is None:
+            continue
+        if should_create is True:
+            instances_to_create.append(instance)
+        else:
+            instances_to_update.append(instance)
+    bulk_create_permission_instances(instances_to_create)
+    bulk_update_permission_instances(instances_to_update)
 
 
 @app.task