+++ concrete_datastore/concrete/management/commands/init_app.py
@@ -11,6 +11,8 @@ from concrete_datastore.api.v1.views import (
 )
 from concrete_datastore.api.v1.permissions import (
     check_instance_permissions_per_user,
+    bulk_create_permission_instances,
+    bulk_update_permission_instances,
 )
 from concrete_datastore.concrete.models import (
     DIVIDER_MODEL,
@@ -82,10 +84,19 @@ def setup_versions_and_permissions():
         app_name='datamodel', version=datamodel_version
     )
     if datamodel_changed is True:
+        instances_to_create = []
+        instances_to_update = []
         for user in get_user_model().objects.filter(
             is_active=True, admin=False, is_superuser=False
         ):
-            check_instance_permissions_per_user(user=user)
+            partial_instances_to_create, partial_instances_to_update = check_instance_permissions_per_user(
+                user=user
+            )
+            instances_to_create.extend(partial_instances_to_create)
+            instances_to_update.extend(partial_instances_to_update)
+
+    bulk_create_permission_instances(instances_to_create)
+    bulk_update_permission_instances(instances_to_update)
 
 
 class Command(BaseCommand):